refactor: rename reputation helpers to utils
